moved autocrop from pp to pl
