+++ spatial_data/pl/utils.py
@@ -2,8 +2,11 @@ from typing import List
 
 import numpy as np
 from matplotlib.colors import LinearSegmentedColormap, ListedColormap
-from skimage.segmentation import find_boundaries
+from skimage.measure import label, regionprops
+from skimage.morphology import closing, square
+from skimage.segmentation import clear_border, find_boundaries
 
+from ..constants import Dims
 from ..pp.utils import _normalize
 
 
@@ -101,3 +104,24 @@ def _label_segmentation_mask(segmentation: np.ndarray, annotations: dict) -> np.
     labeled_segmentation[neg_mask] = 0
 
     return labeled_segmentation
+
+
+def _autocrop(sdata, channel=None, downsample=10):
+    if channel is None:
+        channel = sdata.coords[Dims.CHANNELS].values.tolist()[0]
+    image = sdata.pp[channel].pp.downsample(downsample)._image.values.squeeze()
+
+    bw = closing(image > np.quantile(image, 0.8), square(20))
+    cleared = clear_border(bw)
+    label_image = label(cleared)
+    props = regionprops(label_image)
+    if len(props) == 0:
+        maxr, maxc = image.shape
+        minr, minc = 0, 0
+        downsample = 1
+    else:
+        max_idx = np.argmax([p.area for p in props])
+        region = props[max_idx]
+        minr, minc, maxr, maxc = region.bbox
+
+    return slice(downsample * minc, downsample * maxc), slice(downsample * minr, downsample * maxr)
